requirements.txt
# Core Django and API dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database and storage
psycopg2-binary==2.9.7
django-environ==0.11.2

# AI and external APIs
openai==1.3.3
requests==2.31.0

# Firebase integration
firebase-admin==6.2.0

# Excel export functionality
openpyxl==3.1.2
django-import-export==3.3.1

# Authentication and security
djangorestframework-simplejwt==5.3.0
django-filter==23.3

# Development tools
python-dotenv==1.0.0

# Production server
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-django==4.5.2